flag to stash current version in compiled form

+++ script/cmp_asm.py
@@ -1,3 +1,4 @@
+import argparse
 import filecmp
 import glob
 import os
@@ -5,60 +6,60 @@ import re
 import subprocess
 import tempfile
 
-
-def build(src):
-    cmd = (
-        "C:/Program Files/LLVM/bin/clang-cl",
-        "-Fa",
-        "-IC:/mpir",
-        "-O2",
-        "-Wimplicit-fallthrough",
-        "-Wno-assume",
-        "-Wno-deprecated-declarations",
-        "-Wno-switch",
-        "-c",
-        "-std:c++17",
-        src,
-    )
-    subprocess.check_call(cmd)
-
-
-def simplify(file):
-    cmd = "undname", file
-    s = subprocess.check_output(cmd, encoding="utf-8")
-    s = re.sub(r"A0x\w+", "_", s)
-    open(file, "w", newline="\n").write(s)
-
+parser = argparse.ArgumentParser()
+parser.add_argument("-s", "--save", help="save reference version", action="store_true")
+args = parser.parse_args()
 
 here = os.path.dirname(os.path.realpath(__file__))
+src = os.path.join(here, "..", "src", "*.cc")
+
+# compile the current version to assembly
+cmd = (
+    "C:/Program Files/LLVM/bin/clang-cl",
+    "-Fa",
+    "-IC:/mpir",
+    "-O2",
+    "-Wimplicit-fallthrough",
+    "-Wno-assume",
+    "-Wno-deprecated-declarations",
+    "-Wno-switch",
+    "-c",
+    "-std:c++17",
+    src,
+)
+subprocess.check_call(cmd)
+
 
 # get the current list of source files
 v = []
-for s in glob.glob(os.path.join(here, "..", "src", "*.cc")):
-    s = os.path.basename(s)
-    s = os.path.splitext(s)[0]
-    v.append(s)
+for f in glob.glob(src):
+    f = os.path.basename(f)
+    f = os.path.splitext(f)[0]
+    v.append(f)
 
-# compile the current version to assembly
-build(os.path.join(here, "..", "src", "*.cc"))
-for s in v:
-    simplify(s + ".asm")
-    os.replace(s + ".asm", s + "1.asm")
+# save canonical assembly
+if args.save:
+    ver = "0"
+else:
+    ver = "1"
+for f in v:
+    cmd = "undname", f + ".asm"
+    s = subprocess.check_output(cmd, encoding="utf-8")
+    s = re.sub(r"A0x\w+", "_", s)
+    open(f + ver + ".asm", "w", newline="\n").write(s)
 
-# compile the previous version to assembly
-build(os.path.join(tempfile.gettempdir(), "0", "*.cc"))
-for s in v:
-    simplify(s + ".asm")
-    os.replace(s + ".asm", s + "0.asm")
+if args.save:
+    exit(0)
+
 
 # get the object files out of the way
 for s in v:
     os.remove(s + ".obj")
 
 # compare the assembly output
-for s in v:
-    file0 = s + "0.asm"
-    file1 = s + "1.asm"
+for f in v:
+    file0 = f + "0.asm"
+    file1 = f + "1.asm"
     if not filecmp.cmp(file0, file1):
-        print(s)
-        os.system(f"fc {file0} {file1} > {s+'.diff.asm'}")
+        print(f)
+        os.system(f"fc {file0} {file1} > {f+'.diff.asm'}")
